requirements.txt
dedalus-labs==0.1.1
flask
httpx==0.27.0
importlib-metadata==8.0.0
inflect==7.3.1
jaraco-functools==4.3.0
jaraco.collections==5.1.0
packaging==24.2
pip-chill==1.0.3
platformdirs==4.2.2
python-dotenv==1.2.1
tomli==2.0.1
